azure_tests/pyproject.toml
[tool.poetry]
name = "azure_tests"
version = "0.1.0"
description = "Serverless Azure solution to push message to the queue, read from it and notify via email"
authors = ["Konstantin Mokhov <Konstantin_Mokhov@epam.com>"]
packages = [
    {include= "functions"}
]

[tool.poetry.dependencies]
python = "^3.8"
azure-functions = "^1.7.2"
pytest = "^6.2.4"
pytest-asyncio = "^0.15.1"
pydantic = "^1.8.2"

[tool.poetry.dev-dependencies]

[build-system]
requires = ["poetry-core>=1.0.0"]
build-backend = "poetry.core.masonry.api"
